+++ lib/convert.dart
@@ -3,7 +3,7 @@
 // Author: tomyeh
 library rikulo_convert;
 
-import "dart:async" show Future, Stream;
+import "dart:async";
 import "dart:convert";
 import "dart:io";
 
@@ -14,11 +14,11 @@ import "dart:io";
 /// If specified and the input is more than allowed, [PayloadException]
 /// will be thrown.
 Future<String> readAsString(Stream<List<int>> stream, 
-    {Encoding encoding: utf8, int maxLength}) async {
+    {Encoding encoding: utf8, int? maxLength}) async {
   final result = <int>[];
   await for (final data in stream) {
     if (maxLength != null && (data.length + result.length) > maxLength)
-      throw PayloadException("Too large");
+      throw PayloadException("Over $maxLength (${data.length} + ${result.length})");
     result.addAll(data);
   }
   return encoding.decode(result);
@@ -32,7 +32,7 @@ Future<String> readAsString(Stream<List<int>> stream,
 /// If specified and the input is more than allowed, [PayloadException]
 /// will be thrown.
 Future<dynamic> readAsJson(Stream<List<int>> stream,
-    {Encoding encoding: utf8, int maxLength}) async
+    {Encoding encoding: utf8, int? maxLength}) async
 => json.decode(await readAsString(stream,
       encoding: encoding, maxLength: maxLength));